シンプルにリファクタリング
